requirements.txt
annotated-types==0.7.0
apache-beam==2.60.0
attrs==24.2.0
backports.tarfile==1.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==2.2.1
crcmod==1.7
Deprecated==1.2.15
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
fastavro==1.9.7
fasteners==0.19
google-api-core==2.23.0
google-apitools==0.5.31
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.72.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-bigtable==2.27.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.1
google-cloud-dlp==3.25.1
google-cloud-firestore==2.19.0
google-cloud-language==2.15.1
google-cloud-pubsub==2.27.1
google-cloud-pubsublite==1.11.1
google-cloud-recommendations-ai==0.10.14
google-cloud-resource-manager==1.13.1
google-cloud-spanner==3.50.1
google-cloud-storage==2.18.2
google-cloud-videointelligence==2.14.1
google-cloud-vision==3.8.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.64.0
grpcio-status==1.62.2
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
keyrings.google-artifactregistry-auth==1.1.2
more-itertools==10.5.0
numpy==1.26.4
oauth2client==4.1.3
objsize==0.7.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
proto-plus==1.25.0
protobuf==4.25.3
psycopg2==2.9.10
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
rsa==4.9
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
zipp==3.21.0
zstandard==0.23.0